refactor: Table and Spinner
