+++ packages/components/spinner/src/Spinner.tsx
@@ -1,4 +1,4 @@
-import { cx, css, keyframes } from 'emotion';
+import { cx, css } from 'emotion';
 import React, { forwardRef } from 'react';
 import tokens from '@contentful/f36-tokens';
 import { Box } from '@contentful/f36-core';
@@ -8,251 +8,15 @@ import type {
   PolymorphicComponentProps,
   PolymorphicComponentWithRef,
 } from '@contentful/f36-core';
+import type { SpinnerSize, SpinnerVariant } from './types';
+import { getStyles } from './Spinner.styles';
 
 const DEFAULT_TAG = 'div';
-
-export type SpinnerVariant = 'primary' | 'default' | 'white';
-
-export type SpinnerSize = 'large' | 'medium' | 'small';
-
-const sizes: { [key in SpinnerSize]: string } = {
-  large: '36px',
-  medium: '20px',
-  small: '14px',
-};
 
 const variants: { [key in SpinnerVariant]: string } = {
   default: tokens.gray700,
   primary: tokens.colorPrimary,
   white: tokens.colorWhite,
-};
-
-const animations = {
-  scale1: keyframes`
-  0% {
-    transform: scale(1, 1);
-  }
-  6.666667% {
-    transform: scale(1.5, 0.5);
-  }
-  13.333333% {
-    transform: scale(1, 1);
-  }
-  26.666667% {
-    transform: scale(1, 1);
-  }
-  33.333333% {
-    transform: scale(1, 1);
-  }
-  40% {
-    transform: scale(1, 1);
-  }
-  53.333333% {
-    transform: scale(1, 1);
-  }
-  60% {
-    transform: scale(1.5, 0.5);
-  }
-  66.666667% {
-    transform: scale(1, 1);
-  }
-  100% {
-    transform: scale(1, 1);
-  }
-`,
-  scale2: keyframes`
-  0% {
-    transform: scale(1, 1);
-  }
-  13.333333% {
-    transform: scale(1, 1);
-  }
-  20% {
-    transform: scale(1.5, 0.5);
-  }
-  26.666667% {
-    transform: scale(1, 1);
-  }
-  40% {
-    transform: scale(1, 1);
-  }
-  46.666667% {
-    transform: scale(1, 1);
-  }
-  53.333333% {
-    transform: scale(1, 1);
-  }
-  66.666667% {
-    transform: scale(1, 1);
-  }
-  73.333333% {
-    transform: scale(1.5, 0.5);
-  }
-  80% {
-    transform: scale(1, 1);
-  }
-  100% {
-    transform: scale(1, 1);
-  }
-`,
-  scale3: keyframes`
-  0% {
-    transform: scale(1, 1);
-  }
-  26.666667% {
-    transform: scale(1, 1);
-  }
-  33.333333% {
-    transform: scale(1.5, 0.5);
-  }
-  40% {
-    transform: scale(1, 1);
-  }
-  53.333333% {
-    transform: scale(1, 1);
-  }
-  60% {
-    transform: scale(1, 1);
-  }
-  66.666667% {
-    transform: scale(1, 1);
-  }
-  80% {
-    transform: scale(1, 1);
-  }
-  86.666667% {
-    transform: scale(1.5, 0.5);
-  }
-  93.333333% {
-    transform: scale(1, 1);
-  }
-  100% {
-    transform: scale(1, 1);
-  }
-`,
-  translate1: keyframes`
-  0% {
-    transform: translate(10px, 33.528168px);
-  }
-  6.666667% {
-    transform: translate(10px, 41.764084px);
-  }
-  13.333333% {
-    transform: translate(10px, 33.528168px);
-  }
-  26.666667% {
-    transform: translate(10px, -2.651608px);
-  }
-  33.333333% {
-    transform: translate(10px, -4.471832px);
-  }
-  40% {
-    transform: translate(10px, -2.651608px);
-  }
-  53.333333% {
-    transform: translate(10px, 33.528168px);
-  }
-  60% {
-    transform: translate(10px, 41.764084px);
-  }
-  66.666667% {
-    transform: translate(10px, 33.528168px);
-  }
-  100% {
-    transform: translate(10px, 33.528168px);
-  }
-`,
-  translate2: keyframes`
-  0% {
-    transform: translate(30px, 33.528168px);
-  }
-  13.333333% {
-    transform: translate(30px, 33.528168px);
-  }
-  20% {
-    transform: translate(30px, 41.764084px);
-  }
-  26.666667% {
-    transform: translate(30px, 33.528168px);
-  }
-  40% {
-    transform: translate(30px, -2.651608px);
-  }
-  46.666667% {
-    transform: translate(30px, -4.471832px);
-  }
-  53.333333% {
-    transform: translate(30px, -2.651608px);
-  }
-  66.666667% {
-    transform: translate(30px, 33.528168px);
-  }
-  73.333333% {
-    transform: translate(30px, 41.764084px);
-  }
-  80% {
-    transform: translate(30px, 33.528168px);
-  }
-  100% {
-    transform: translate(30px, 33.528168px);
-  }
-`,
-  translate3: keyframes`
-  0% {
-    transform: translate(50px, 33.528168px);
-  }
-  26.666667% {
-    transform: translate(50px, 33.528168px);
-  }
-  33.333333% {
-    transform: translate(50px, 41.764084px);
-  }
-  40% {
-    transform: translate(50px, 33.528168px);
-  }
-  53.333333% {
-    transform: translate(50px, -2.651608px);
-  }
-  60% {
-    transform: translate(50px, -4.471832px);
-  }
-  66.666667% {
-    transform: translate(50px, -2.651608px);
-  }
-  80% {
-    transform: translate(50px, 33.528168px);
-  }
-  86.666667% {
-    transform: translate(50px, 41.764084px);
-  }
-  93.333333% {
-    transform: translate(50px, 33.528168px);
-  }
-  100% {
-    transform: translate(50px, 33.528168px);
-  }
-`,
-};
-
-const styles = {
-  circle1Scale: css({
-    animation: `${animations.scale1} 1s linear infinite normal forwards;`,
-  }),
-  circle2Scale: css({
-    animation: `${animations.scale2} 1s linear infinite normal forwards;`,
-  }),
-  circle3Scale: css({
-    animation: `${animations.scale3} 1s linear infinite normal forwards;`,
-  }),
-  circle1Translate: css({
-    animation: `${animations.translate1} 1s linear infinite normal forwards;`,
-  }),
-  circle2Translate: css({
-    animation: `${animations.translate2} 1s linear infinite normal forwards;`,
-  }),
-  circle3Translate: css({
-    animation: `${animations.translate3} 1s linear infinite normal forwards;`,
-  }),
 };
 
 export type SpinnerInternalProps = {
@@ -290,19 +54,13 @@ export const _Spinner: PolymorphicComponentWithRef<
   },
   forwardedRef,
 ) => {
+  const styles = getStyles();
   return (
     <Box
       as={DEFAULT_TAG}
       display="inline-block"
       {...otherProps}
-      className={cx(
-        css({
-          height: customSize ? `${customSize}px` : undefined,
-          verticalAlign: 'middle',
-          width: customSize ? `${customSize}px` : sizes[size],
-        }),
-        className,
-      )}
+      className={cx(styles.root({ size, customSize }), className)}
       ref={forwardedRef}
       testId={testId}
     >
