+++ packages/components/spinner/src/types.ts
@@ -1,0 +1,3 @@
+export type SpinnerVariant = 'primary' | 'default' | 'white';
+
+export type SpinnerSize = 'large' | 'medium' | 'small';
