+++ packages/components/table/src/Table.tsx
@@ -4,26 +4,28 @@ import { Box } from '@contentful/f36-core';
 import type { BoxProps } from '@contentful/f36-core';
 import tokens from '@contentful/f36-tokens';
 
-const styles = {
-  inline: css({
-    borderRadius: tokens.borderRadiusMedium,
-    boxShadow: `0 0 0 1px ${tokens.gray200}`,
-    'th:first-child': {
-      borderTopLeftRadius: tokens.borderRadiusMedium,
-    },
-    'th:last-child': {
-      borderTopRightRadius: tokens.borderRadiusMedium,
-    },
-    'tr:last-child td:first-child': {
-      borderBottomLeftRadius: tokens.borderRadiusMedium,
-    },
-    'tr:last-child td:last-child': {
-      borderBottomRightRadius: tokens.borderRadiusMedium,
-    },
-  }),
-  root: css({
-    width: '100%',
-  }),
+const getStyles = () => {
+  return {
+    inline: css({
+      borderRadius: tokens.borderRadiusMedium,
+      boxShadow: `0 0 0 1px ${tokens.gray200}`,
+      'th:first-child': {
+        borderTopLeftRadius: tokens.borderRadiusMedium,
+      },
+      'th:last-child': {
+        borderTopRightRadius: tokens.borderRadiusMedium,
+      },
+      'tr:last-child td:first-child': {
+        borderBottomLeftRadius: tokens.borderRadiusMedium,
+      },
+      'tr:last-child td:last-child': {
+        borderBottomRightRadius: tokens.borderRadiusMedium,
+      },
+    }),
+    root: css({
+      width: '100%',
+    }),
+  };
 };
 
 export type TableInternalProps = Omit<BoxProps<'table'>, 'as'> & {
@@ -43,6 +45,7 @@ export const Table = forwardRef<HTMLTableElement, TableInternalProps>(
     },
     forwardedRef,
   ) {
+    const styles = getStyles();
     return (
       <Box
         cellPadding="0"
