+++ packages/components/table/src/TableRow.tsx
@@ -5,23 +5,25 @@ import tokens from '@contentful/f36-tokens';
 import { Box } from '@contentful/f36-core';
 import type { CommonProps } from '@contentful/f36-core';
 
-const styles = {
-  root: css({
-    '&:last-child td': {
-      borderBottom: 'none',
-    },
-    '&:hover td': {
-      backgroundColor: tokens.gray100,
-    },
-  }),
-  selected: css({
-    'td, th': {
-      backgroundColor: tokens.blue100,
-    },
-    'td:last-child': {
-      boxShadow: `inset -2px 0 0 ${tokens.blue500}`,
-    },
-  }),
+const getStyles = () => {
+  return {
+    root: css({
+      '&:last-child td': {
+        borderBottom: 'none',
+      },
+      '&:hover td': {
+        backgroundColor: tokens.gray100,
+      },
+    }),
+    selected: css({
+      'td, th': {
+        backgroundColor: tokens.blue100,
+      },
+      'td:last-child': {
+        boxShadow: `inset -2px 0 0 ${tokens.blue500}`,
+      },
+    }),
+  };
 };
 
 export type TableRowInternalProps = HTMLProps<HTMLTableRowElement> &
@@ -43,6 +45,7 @@ export const TableRow = forwardRef<HTMLTableRowElement, TableRowProps>(
     },
     forwardedRef,
   ) {
+    const styles = getStyles();
     return (
       <Box
         {...otherProps}
